styled cards, combined files, refactored imports
