+++ src/components/Card/CardGrid.js
@@ -1,18 +1,19 @@
 import React, { useContext } from 'react';
 import { Grid, ResponsiveContext } from 'grommet';
 
-export const CardGrid = ({ children, ...rest }) => {
+const CardGrid = ({ children, ...rest }) => {
   const size = useContext(ResponsiveContext);
   return (
     <Grid
-      columns={size !== 'small' ? { count: 2, size: 'medium' } : '100%'}
-      rows={['auto', 'flex']}
+      columns={size !== 'small' ? { count: 2, size: 'auto' } : '100%'}
+      rows={['auto']}
       gap="large"
       justify="center"
       {...rest}
-      fill
     >
       {children}
     </Grid>
   );
 };
+
+export default CardGrid;
