--- src/components/Cards/index.js
@@ -1,6 +1,0 @@
-export * from './ContentCard';
-export * from './CommunityImage';
-export * from './Card';
-export * from './CardGrid';
-export * from './CardImage';
-export * from './ScheduleCards';
