+++ src/components/index.js
@@ -6,14 +6,18 @@ import HPEDevFooter from './HPEDevFooter';
 import Layout from './Layout';
 import Leaderboard from './Leaderboard';
 import SideNav from './SideNav';
+import { Card, CardGrid, ScheduleCard } from './Card';
 
 export {
   ButtonSplit,
+  Card,
+  CardGrid,
   DevLogo,
   Footer,
   Header,
   HPEDevFooter,
   Layout,
   Leaderboard,
+  ScheduleCard,
   SideNav,
 };
